+++ config/twig.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use Globetrotters\AiPresenceBundle\Cache\ArtefactCache;
+use Globetrotters\AiPresenceBundle\Serving\BreadcrumbRenderer;
 use Globetrotters\AiPresenceBundle\Twig\AiPresenceExtension;
 
 return static function (ContainerConfigurator $container): void {
     $container->services()
         ->set(AiPresenceExtension::class)
-        ->args([service(ArtefactCache::class)])
+        ->args([service(ArtefactCache::class), service(BreadcrumbRenderer::class)])
         ->tag('twig.extension');
 };